Update patch set 2
Patch Set 2:

The code seems fine, but I just realized we are missing a crucial component to this: the actual TermIndex has to become repo-aware. I filed T151053 for that. 

This change is fine to merge I think, but I would prefer to first see the implementation of T151053, in case we have to change the interface a bit.

Patch-set: 2
Label: Code-Review=0