Drop public setters from TermIndexEntry
In most cases those setters could be replaced by passing a relevant
value as an array element to TermIndexEntry's constructor.
The only actual use of the setter was in unit test, while asserting
that modifying the cloned instance of TermIndexEntry does not affect
the "original" instance. As there seem to be no need for TermIndexEntry
to be mutable, such test is no longer needed too.

Change-Id: Iebb76a1c3006f0ae339b5b2f12beb5a89c7e1d5c