Update patch set 1
Patch Set 1:

As I tried to explain the weight is not entirely unused. It is calculated, stored, and used for ordering. But all this code is in the TermSqlIndex class. I could not find other code that does anything with the weight.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0