Split InterwikiSorting hook handlers from ParserOutputUpdateHookHandlers
In preparation to remove these and use InterwikiSorting extension
instead.

The new hook handlers also work with the new settings that
InterwikiSorting has, as well as the old wikibase configs.

Next steps are to add config to mediawiki-config (T111023)
and then after this code is deployed, then we can change
the config variables used there to the new ones and
remove support for the old configurations.

Bug: T151652
Change-Id: I5b6fd2602f5d6f08114a5877dfc872b9d5db672c