Allow calling getNumericId on foreign ItemIds and PropertyIds
It seems that disallowing calling getNumericId on foreign ItemId
and PropertyId instances (ie. containing a prefix in their serializations)
was not right.
Although it definitely it should not be assumed that "Q1" and "foo:Q1"
refer to the same entity of type "item" and numeric part of id "1",
the actual code accessing data from the database (ie. accessing
numeric values from the DB) must make use of getNumericId to get
the numeric part of both "Q1" and "foo:Q1".

It should be the responsibility of the caller to make sure the
right database is accessed and that the number part of id makes
sense in the given context. ItemId/PropertyId should not really
know anything about this.