Update patch set 3
Patch Set 3:

I'm not sure how it's different from what the second part of the test (second snak) is doing. It's using the value, which is already in the bag, and thus the value is not output but the clause is. That's exactly the code path that is taken when the value is duplicated. Note again, it's not the same snak twice, it's different snaks using the same value - exactly the code path you describe.

Patch-set: 3