Update all Wikibase\Repo\Tests\… folders (but Api) to PSR-4
I'm skipping the Api folder for now because it is big and a class is used
outside of Wikibase. The rest is not used anywhere else. The basic idea is:
No production code can depend on test code. If this change passes CI it
must be fine and can be merged.

Change-Id: I679ece9059530e0ad1d221cfbd1cfbaa724b3129