Update patch set 8
Patch Set 8:

Yes. Parser contains hundreds of things this code does not need. Passing only two of them is incredibly better than passing the entirety of Parser.

Now I wonder why so many places in the code need to be changed to pass these two parameters around? It's really weird to add this to a method called "parse". What would be a more "correct" solution for this?

Patch-set: 8
Label: Verified=0