Catch missing exception in SpecialGoToLinkedPage
Have a look at
https://www.wikidata.org/wiki/Special:GoToLinkedPage/enwiki/x%3AQ1
This is an exception that is shown to the user when he enters something
like "Item:Q1". What happens? The "Item:" prefix is interpreted as the name
of a foreign repository.

I followed all code paths but can not find one that throws an InvalidValue…
exception.

The additional change in RevisionBasedEntityLookup is only losely related.
I found this while debugging this error.

Change-Id: Ie8f02e7a90a986656a67665b6afed938aa69ade5