Make SiteLinkConflictLookup return ItemIds instead of numbers
This does not get rid of the fact that Item IDs are numeric, but this
patch moves this fact into the (only) implementation. It's much closer
to the actual SQL query now. It's an implementation detail now.

This is related to all the other attempts in the past years and weeks
to get rid of numeric IDs.

Change-Id: I4f79544f1318dd32509611fa2fb64bf5d0c0729e