Avoid term box jumping down while the page loads
This reserves the exact space the toggler on top of the term box will need.
It's 1 line-height (1.6em) plus the margin-bottom of the aliases section
(1em). This is necesarry because margins collapse.

The moment the toggler is added to the DOM, the second rule applies and
removes the space again.

This additional space will also be visible on diff and old revision pages.

Change-Id: I77fcdbe2b8cb7c3d6e31f7ffa0dd9e1043b4ed85