Update patch set 1
Patch Set 1:

Jumping where, and how much? There are multiple places where stuff jumps. This patch is about the expand button on *top* of the terms table. There is an other button *below* the terms table, but this button is not always there. This patch also does not change the situation when the term box is collapsed. Stuff will still jump then. This patch does have an effect only when the term box is expanded (which is the default for a while now).

Patch-set: 1