Update patch set 1
Patch Set 1:

"Not working" is probably the least helpful thing to say. What do you mean? What happens? Are you looking at the right place? Which browser? Are you logged in? Is a Babel box set up?

I found a flaw when not being logged in and fixed it in PS2.

Patch-set: 1
Label: Verified=0