Fix language code returned by Lua's mw.wikibase.getLabelWithLang
My first attempt was to add an "if ( $term instanceof TermFallback )", and
to extract this to a private "getActualLanguageCode" method, because it is
needed two times for the label and the description. But code with such a
feature check is ugly and just not necesarry in this case. Yes, the class now
binds against a concrete implementation instead of an interface. Nothing wrong
with this. The interface was not there because of dependency inversion. The
only benefit it had was that it was possibly to replace the implementation
with an other one. But this was not used, and therefor "dead code". It's still
possible to change this code, because all the code I'm touching here lives in
the same code base anyway.

Bug: T152241
Change-Id: I1ba2142d7282a9e92252c825740553b24223fc4a