Apply property ordering in StatementGroupListView
This patch introduces a little bit of duplicate code in two places:
The top level factory code ("new WikiPagePropertyOrderProvider" and so
on) is partly duplicated in client and repo.
And the actual ordering is duplicated (see Scribunto_LuaWikibaseLibrary
for the duplication).
Both cases should be refactored later, in my opinion.

Bug: T150788
Change-Id: I979cedf94ff5053d24d2aad333c707074f20c324