Remove/fix incomplete PHPDoc tags and remove misplaced commas
The comma is a copy paste mistake that happened in I979cedf.

We found that it is best practice to *not* have @param documentation
when there is a @dataProvider. This is often useless (e.g. when an
"invalid argument" provider lists many different types), error prone,
and of not much value.

Change-Id: Ia6597edb7ea9ce56cd2b5e863a7274c7c987234e