Update patch set 2
Patch Set 2: Code-Review-1

This morning I was told that this patch was mentioned in the Scrum of Scrums for being a "blocker". I'm very sorry to hear that, but I have to ask what we (the Wikidata team at WMDE) are supposed to do here? Why is this marked as "WIP"? "WIP" usually means "this is not ready for review". Is there a Phabricator ticket, and can you please add it to the tasks description, so we can add it to our sprint?

-1 to get your attention, and because of the WIP.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1