Allow optional usage of Client's data retrieval services in the Repo component
This allows WikibaseRepo to optionally get an instance of
EntityDataRetrievalServiceFactory from WikibaseClient instance,
and use (dispatching) services configured in the client.

NOTE: this makes Repo component optionally depend on Client.
This is generally wrong, and only introduced as a temporary
solution in order to allow Repo to access entity data from
foreign repositories. This would be fixed once
DispatchingServiceFactory and RepositoryServiceContainer
in particular do not rely on the instance of WikibaseClient
to provide services.

Also, as DispatchingServiceFactory is an implementation
of EntityStoreWatcher, it could be used to properly purge
entity meta data cache used by WikibaseRepo. When entity
is updated or deleted, DispatchingServiceFactory passes
the cache purge request to the prefetching
EntityMetaDataAccessor instance configured for the repository
the entity belongs to.

Bug: T155623
Change-Id: I15d33dd688a0eb1de684d963ae94640aa3509f6d