Allow optional usage of Client's data retrieval services in the Repo component
This allows WikibaseRepo to optionally get an instance of
EntityDataRetrievalServiceFactory from WikibaseClient instance,
and use (dispatching) services configured in the client.

NOTE: this makes Repo component optionally depend on Client.
This is generally wrong, and only introduced as a temporary
solution in order to allow Repo to access entity data from
foreign repositories. This would be fixed once
DispatchingServiceFactory and RepositoryServiceContainer
in particular do not rely on the instance of WikibaseClient
to provide services.

Service instantiatiors defined in RepositoryServiceWiring file
now also define "WikiPageEntityMetaDataAccessor" service
that provides a meta data accessor for Container's
EntityRevisionLookup (and in turn also for
the DispatchingEntityRevisionLookup). This accessor
currently implement the EntityStoreWatcher interface and gets
notified when entity change event happens, so the prefetcher's
cache is updated according to changes.

Bug: T155623
Change-Id: I15d33dd688a0eb1de684d963ae94640aa3509f6d