Allow optional usage of data retrieval services factory in repo component
This introduces an interface for a factory of services used to be
provide instances of data retrieval/lookup services that are configured
accordingly to repository settings (including foreign repositories).

DispatchingServiceFactory is implementing this interface, providing
dispatching wrappers around lookup services.

This also allows WikibaseRepo to optionally get an instance of
DataRetrievalServiceFactory from WikibaseClient instance, and use
(dispatching) services used by client.

Also, this makes DispatchingServiceFactory an implementation
of EntityStoreWatcher, so it could be used to properly purge
cache used by WikibaseRepo. When entity is updated or deleted,
DispatchingServiceFactory passes the cache purge request
to the EntityPrefetcher instance configured for the repository
entity belongs to.

TODOs:
 - test EntityStoreWatcher stuff in DispatchingServiceFactory
 - split into more patches (interface, Watcher thing)
 - some better tests for the case of using DispatchingServiceFactory

Change-Id: I15d33dd688a0eb1de684d963ae94640aa3509f6d