Allow optional usage of entity data retrieval services factory in repo component
This allows WikibaseRepo to optionally get an instance of
EntityDataRetrievalServiceFactory from WikibaseClient instance,
and use (dispatching) services configured in the client.

Also, as DispatchingServiceFactory is an implementation
of EntityStoreWatcher, it could be used to properly purge
entity meta data cache used by WikibaseRepo. When entity
is updated or deleted, DispatchingServiceFactory passes
the cache purge request to the prefetching
EntityMetaDataAccessor instance configured for the repository
the entity belongs to.

TODOs:
 - some better tests for the case of using DispatchingServiceFactory

Change-Id: I15d33dd688a0eb1de684d963ae94640aa3509f6d