Update patch set 5
Patch Set 5: Code-Review+2

I realized that the issue of having a hard coded list of entity types is not exclusive to this patch. The previous patch introduces it. This patch here copies it to a place where it's not obvious any more why this is done.

All these places should use the same configuration, and not copy hard coded values.

My follow up is not perfect because it can possibly (only on third party installations with custom entity types) lead to a situation where the special page offers entity types that are not indexed in the terms table. I still believe this is better than duplicating this hard coded list of entity types.

Patch-set: 5
Label: Code-Review=+2