Update patch set 5
Patch Set 5:

The change you are mentioning is a result of the discussion we had. assertTag was previously testing way to many things that are not relevant. If something is wrapped in a <div> or not is not relevant to make sure the form can be used and does what it is supposed to do. We never had custom CSS/JS in mind when we wrote this. Alex removed this because I asked him to do so.

Patch-set: 5