Update patch set 2
Patch Set 2: Code-Review-1

It appears the commit message is missing the WIP tag. Just adding two slashes in from of a line of code is not a proper bugfix. I suggest to either remove it entirely, or replace it with a more proper log or warning method, or fix the CI setup that is obviously incomplete (the sites table does not contain the site the tests are run with).

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1