Update patch set 8
Patch Set 8: Code-Review+2

(1 comment)

I still do not like that this patch introduces an implementation of EntityTitleLookup that is incompatible with some of the existing callers of that interface. But since we seem to agree on what the design should look like in the end, and there's only disagreement on how to get there, let's start moving.

Patch-set: 8
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2