Tests: Add users to DB when they are given advanced rights
In general, it doesn't make sense to assign user groups to users who
are not in the database, as it will create bogus database entries
(rows in the user_groups table with ug_user = 0). For tests, this is
not so much of an issue, as the database doesn't need to be
consistent. But I plan to start enforcing that the user ID > 0 when 
assigning users to a group (I93c955dc7a970f78e32aa503c01c67da30971d1a)
so all these tests would fail in that scenario without this change.

Change-Id: Ia9616e1e35184fed9058d2d39afbe1038f56d7fa