Use default wiki id in example settings.
Until now, we force the local wiki ID to "mywiki" in the exampel settings.
This triggers a warning if "mywiki" is not known to the SiteLookup.

Instead of forcing a fixed value, we can just rely on the default for the
local site ID, which is the wiki's database name. That is consistent with
the conventions used on the WMF cluster.

Bug: T153729
Change-Id: Ia0d578a8215280592001cfa5f413fb76e15a2d2f
(cherry picked from commit 8e5d3af54d8c5da0783ce1427c8deaf2d1a5d284)
(cherry picked from commit 166c4fc5b71f691c5b78579a4ab1ce69d940b777)