Make EntityDataRetrievalServiceFactory provide TermSearchInteractor
Adds TermSearchInteractorFactory interface for factories
creating instances of TermSearchInteractor configured to use
a given display language.

Two implementations of the interface are added:
 - TermIndexSearchInteractorFactory providing interactors configured
   to use a single repository,
 - DispatchingTermSearchInteractoryFactory providing interactor
   dispatching (routing) the search to the right repository based
   on the type of the entity.

WikibaseRepo now (optionally) gets the
DispatchingTermSearchInteractorFactory instance
from the EntityDataRetrievalServiceFactory and uses it
to return the right interactor in WikibaseRepo::newTermSearchInteractor.

This also introduces a ConfigurableTermSearchInteractor interface
for TermSearchInteractors that could be configured using
TermSearchOptions (as TermIndexSearchInteractor already was doing).
DispatchingTermSearchInteractor implements the new interface,
and passes current search options to the interactor actually
used in the given term search.

Bug: T153793
Change-Id: I9212885005d4753661996acdac607dc4f88eb21d