Make EntityDataRetrievalServiceFactory provide TermSearchInteractor
Adds TermSearchInteractorFactory interface for factories
creating instances of TermSearchInteractor configured to use
a given display language.

Two implementations of the interface are added:
 - TermIndexSearchInteractorFactory providing interactors configured
   to use a single repository,
 - DispatchingTermSearchInteractoryFactory providing interactor
   dispatching (routing) the search to the right repository based
   on the type of the entity.

WikibaseClient uses DispatchingTermSearchInteractorFactory to
return the search interactor in its newTermSearchInteractor
metchod.

Also WikibaseRepo now (optionally) gets the
DispatchingTermSearchInteractorFactory instance
from the EntityDataRetrievalServiceFactory and uses it
to return the right interactor in WikibaseRepo::newTermSearchInteractor.

This change also defines a service named PrefetchingTermLookup
in RepositoryServiceWiring file. This makes both DispatchingTermBuffer
and single-repo TermSearchInteractoryFactory use single instance
of the PrefetchingTermLookup for a single repo.

Bug: T153793
Change-Id: I9212885005d4753661996acdac607dc4f88eb21d