Update patch set 4
Patch Set 4:

Please don't just remove warnings with no replacement, no comment in the code, and no ticket to put it back. This warning exists for a good reason. 

I'm not reverting this patch to not break CI again, but I'm not happy about how this went. Self-merging this kind of thing is ok as a hot fix on the deployment branch, but not on the master branch.

I have filed T153854 to restore the warning. Let's instead fix the configuration problem it is complaining about.

Patch-set: 4