Update patch set 4
Patch Set 4:

> Please don't just remove warnings with no replacement, no comment
 > in the code, and no ticket to put it back. This warning exists for
 > a good reason.

Any suggestions on how to handle it in the future? I had hoped that an UBN task that's open for a full CEST business day would result in the team responsible for the extension writing a proper fix, or if not that then some feedback at least, but that didn't really work out.

Patch-set: 4