Update patch set 4
Patch Set 4:

> Any suggestions on how to handle it in the future? I had hoped that an UBN task that's open for a full CEST business day would result in the team responsible for the extension writing a proper fix, or if not that then some feedback at least, but that didn't really work out.

The only thing I saw was a hack that did nothing but literally patching around an issue we, the Wikidata team, could not have caused. The Phabricator ticket was never assigned and never picket up because there was nothing we could do with the information given. The warning message was there for years. I hope you understand that blaming a years old line of PHP code for breaking something entirely unrelated in a JavaScript test is really, really awkward? The line warns about a misconfiguration, and that's exactly what should have been fixed. It's as simple as that. Keep calm, find out what changed when your test started to break, and fix it.

Patch-set: 4