Update patch set 5
Patch Set 5: Code-Review+1

(1 comment)

Eventually, the logic for constructing ChangeOps from a JSON structure should be factored out of the API module, and be moved into a ChangeOpDeserializer service. But this is OK as a first step. I see no immediate problem with the code. Let's just not forget to clean it up later.

Patch-set: 5
Label: Code-Review=+1