Update patch set 3
Patch Set 3:

Thanks for the additional information. What this means is that properly refactoring this means we have to change the entire error handling in the frontend as well. Considering all this we are talking about three or more devs working on this for at least two weeks, more probably four weeks, because of code that got deprecated literally days (!) ago.

Please understand that neither a WIP patch, nor complaining in the Scrum of Scrums, nor a discussion in an abandoned patch is going to make us invest so much resources for something nobody told us about beforehand. This is exceptionally outside of our current budget plans, even if refactoring exactly this error handling is something we are actually planing to do (most notably T54843 and T153858). But we did not picked this up for the current quarters.

We are discussing this right now in our mailing list and will document results in T153359. Please do so as well and help us getting T153359 in a shape we can collaboratively work on it.

Patch-set: 3