Update patch set 2
Patch Set 2:

Tests run on whatever local MediaWiki installation you run them. This local wiki does have a UI language configured. There is nothing in the PHPUnit bootstrap that enforces the configured language to be English. Tests run with whatever configuration is there, and only override settings if a test explicitly asks for.

For example, a test that is not marked with "@database" will mess with the real database.

You can say that this is "leaking global state into a test". But this shouldn't be the only argument. I found it very helpful to run the same tests with different configurations. This makes sure that the tests are not testing details they should not test. For example, a test should not assume a specific database object is there, just because you have it on your machine when you wrote the test. Same can be said for messages. It's usually not a good idea to test localized messages, because the translation can change every day. Testing for English messages is not as bad, because English is in the same code base and changes are unable to break tests unexpectedly. But the same argument: such a test is probably testing details (specific messages) it should not test. That's why I prefer to use qqx.

I can't tell why Travis is running these tests with a Russian setup. But the fix is correct and very common: when a test compares messages, it must make sure the language is set up correctly.

Patch-set: 2