Update patch set 2
Patch Set 2: Code-Review-1

(5 comments)

The documentation of this interface is very much incomplete. I believe I understand what it is meant to do, because I know an other patch that goes along with this. But I'm afraid a coder not familiar with the rest of the code base will not understand what this interface is meant to do, where and why.

Patch-set: 2
Label: Code-Review=-1