Make UnitConverterTest compatible with non-BC floating point math
This code uses the DecimalMath class, which supports both BC and
floating point arithmetics. The class under test does not care. Why
should the test be limited then?

This is the most minimal patch I could think of to make it compatible
with both modes. This patch adds a bit of complexity because I'm using
regular expressions now.

Change-Id: Ie97e2fca2f55b2390d166ed072ced40a15f4ef14