Update patch set 2
Patch Set 2:

Why was the name changed to EditEntityApiChangeOpDeserializer? How is this specific to the EditEntity module? Do we have a need to support  serializations for change ops that would not be able to use this interface?

To me it seems like this is a very generic interface - raw data structure in, ChangeOps out. No need to know anything about API modules.

Patch-set: 2