Update patch set 2
Patch Set 2:

Exact same discussion as in Id1e5a8f. At the moment this "serialization" format appears to be undocumented, and exclusively used in EditEntity. We can change this, but not by simply changing a name and assuming everything is generic and fine, but by documenting it and that way showing that it is generic.

Patch-set: 2