Update patch set 5
Patch Set 5:

Too late but FWIW, regarding the class name: IMO "trimming" (that Thiemo has initially suggested) would have been better than "trimmed" (that it ended up being like). If I get it right, you could enter a text with leading spaces etc and the field would trim it. The field does not require to get "trimmed" input, it does _trim_ it.

But yeah, names. This class not only trims the text but also normalizes the input., So as Aleksey said, I guess we would rather prefer to have some better name for the class (which I have no suggestions for). So let's rename it to something better when we have an idea!

The general idea behind the patch is good!

Patch-set: 5