Update all composer.json version numbers to "~x.x.x"
See https://getcomposer.org/doc/articles/versions.md#tilde

"~1.0" and "^1.0.0" both mean all 1.0.x bugfix releases *and* all 1.x.0
feature releases are pulled in automatically. This is a risk, because
every release can contain unexpected breaking changes. But what is the
benefit of pulling in feature releases automatically? The feature is not
used. It will become used with a patch that starts using it. That patch
can easily update the version number to "~1.1.0", when necesarry.

The only problem is when a feature release also contains critical
bugfixes. These bugfixes will not be pulled in automatically. I believe
the additional Gerrit patches we need to do in such cases is worth the
better stability.

Change-Id: I497fe7da8bcce0488d11a9c891a118570139e9b1