Update patch set 3
Patch Set 3:

I'm not going to waste more time on this. Realizing how much time we all spend on this is just painful. Especially because the line you are complaining about now was *not* introduced with this patch, and the argument is (even if valid) hypercritical and of no practical consequence. Feel free to upload your own patch, but please do not downvote patches just because you found a potential issue in a file a patch is touching.

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0