Update patch set 10
Patch Set 10: Code-Review-1

(3 comments)

Discussed in a pair review session yesterday. I'm fine with removing the "EditEntityApi" part. Sorry for the fuzz I created.

I had a deeper look into the code. The format of the "$data" array is more or less a "standard" entity serialization, plus a few additional keys here and there (for example, it's possible to request deletions of aliases or even emptying the entire entity). Documenting this is really critical, because some of these details are currently nowhere documented (not in any serializer or deserializer), except in the code where these array keys are detected.

Note that there are other API modules that require the user to paste a JSON serialization and turn them into ChangeOps internalls, e.g. SetClaim. But this is *not* compatible with the JSON serialization in this patch, and I believe we can never make these different API modules use the same "change-op-callbacks".

However, I believe Daniels argument holds: Will there be a need for a different way of creating ChangeOps, from different input? I believe we do not need this. And even if we do, these "change-op-callbacks" will be more specific then, and need a more specific name. The callbacks in this patch here can use the most generic name, because they are for the most generic API module.

-1 because of the issues described in the inline comments.

Patch-set: 10
Label: Code-Review=-1