Update patch set 4
Patch Set 4: Code-Review+1

FWIW: If I read the current code right, the only ChangeOp which is needing the Entity is ChangeOpSiteLink (there is a check if the item has a link with the given site id before creating the ChangeOp).
I am not sure if this is right, but if we need this check there, that would be the argument in favour of Thiemo's opinion.

Not knowing about details of the ChangeOpSiteLink I personally was agreeing with Daniel that the interface should only expect a serialized change op being provided.
Regardless this site link case I brought here I understand concerns Thiemo is raising. I am not sure whether we know there are ChangeOp that _must_ know what the Entity looks like, or is it still hypothetical.

And generally I think we could decide on one or other way know and re-think it once we face problems. That way we could move forward with all other required work with cleaning up the ChangeOp-related code we have in front of us.

Patch-set: 4
Label: Code-Review=+1