Update patch set 5
Patch Set 5:

Discussed with Leszek and Aleksey. Decided we do *not* add the old entity to this interface. If we do this would mean it is possible to apply the exact same change (the user submits a JSON blob representing his edit request) to two different entities, and this results in two different edits. Do we really want to allow this? We believe not. So $oldEntity is *out*.

I hope we agree that this method should return ChangeOp (the interface).

Implementations should be allowed to return null, if the provided array structure is valid, but does not contain anything that represents a change.

But is an array structure really a "serialization", and turning it into ChangeOp objects "deserialization"? We are using the term "serialization" way to often, for way to many things. What the user is actually doing when he adds a "remove" key is requesting a specific edit.

Is the interface a "Mapper" that "maps" user provided data to ChangeOps? "Builder"? "Factory"? "Parser"?

deserialize() vs. getChangeOp() vs. instantiateChangeOp() vs. createChangeOp()?

$serialization vs. $data vs. $entityData vs. $entitySerialization vs. $changeRequest vs. $entityChangeRequest. Personally I find it important to mention "entity" *somewhere*, because this is not a arbitrary data model element (e.g. only a statement), it *always* starts at the entity level.

We are almost ;-) happy with these suggestions. Note that the second does not get the entire array, but only a part of it. The only assumption is that the user provided data always starts with string array keys at the first level.

interface ChangeOpParser {
    public function createEntityChangeOp( array $changeRequest );
}

interface ChangeOpMapper {
    public function isMapperFor( string $elementKey );
    public function createChangeOp( mixed $changeRequest );
}

Patch-set: 5