Add Special:EntityPage for addressing entities without knowing the namespace
Special:EntityPage expects entity ID as a sub page parameter.
The special page uses EntityTitleLookup to get a page related to the entity,
and generates a redirect to this page.
It could be redirecting to entity's page on the local wiki when
given a local entity ID, and to Special:EntityPage on the foreign
repo's wiki when given a foreign entity ID.

This allows wikis to link to an entity page without needing to know
namespace names of entity types configured on the target wiki.

Bug: T153499
Change-Id: I4452689ef7869dd61c158e8a314b564680945967