Add Special:EntityPage for addressing entities without knowing the namespace
Special:EntityPage expected entity ID as a sub page parameter.
Given local entity ID, it redirects to entity's page on the local wiki.
Given foreign entity ID, it redirects to Special:EntityPage on the foreign
repo's wiki.

This allows wikis to link to an entity page without needing to know
namespace names of entity types configured on the target wiki.

Bug: T153499
Change-Id: I4452689ef7869dd61c158e8a314b564680945967