Update patch set 5
Patch Set 5:

There is a detail I discussed further with Aleksey. Our second proposal should look like this:

interface ChangeOpMapper {
    public function isMapperFor( string $elementKey );
    public function createChangeOp( array $changeRequest );
}

Note the array type hint. Extensions should not be allowed to use a string to represent a change, e.g. { "language": "en" }. It should be enforced to be an array, e.g. { "language": { "code": "en" } }, with "language" being the $elementKey and { "code": "en" } being the $changeRequest. This way it's possible to extend this with additional keys, e.g. "delete" or "add", as the current ChangeOps do.

Patch-set: 5
Label: Verified=0