Update patch set 5
Patch Set 5:

I'm happy to see that this discussion is progressing in a constructive way. Just a quick note regarding the usage of "serialization":

We use "serialization"/"deserialization" throughout the wikibase code to mean conversion between nested array structures and objects. I was never happy with that usage, but we at some point compromised on using it, mostly for lack of a good alternative - I once proposed "objectify" and "arrayalize", but nobody liked it :)

So, turning an array structure into a ChangeOp is a "deserialization" according to our convention. I'd be happy to have a better term for this, but if we change it, we should change it everywhere.

Patch-set: 5