Update patch set 3
Patch Set 3:

> What you say is exactly the intention of this patch, as explained in the commit message. What is the benefit of automatically pulling new features in, if they are not used?

It allows people that extend Wikibase to use these new features in these libs / other packages that work alongside Wikibase to use them.

Lets take data-values/serialization as an example.
This patch changes the required version to ~1.1.0 which mean greater than 1.1.0 and less than 1.2.0
In theory if there were a 1.3 and a 1.4 release adding features that wikibase was not yet using then you say we would leave the requirement at ~1.1.0

If someone were to then have an altered / new datavalue or extension that required a feature in data-values/serialization 1.3 or 1.4 they would no longer be able to install their datavalue / extension alongside wikibase.
They would have to come and make a patch to change the required version up to 1.4 OR back to ~1.1 which is what the wikibase package actually requires.

In summary, this change introduces fake requirements / strictness and generally makes working with everything harder.

Again revisiting:

 > What is the benefit of automatically pulling new features in, if they are not used?

Flexibility, Extensibility, Sanity
Wikibase is not a final product, but a building block!

Patch-set: 3