Update patch set 3
Patch Set 3:

And in response to some of the other points in the commit msg.

> This is a risk, because every release can contain unexpected breaking changes.

unexpected breaking changes in the libraries should be fixed in the libraries and released as a bug fix version.

> But what is the benefit of pulling in feature releases automatically?

Flexibility & Extensibility (see previous gerrit comment)

> That patch can easily update the version number to "~1.1.0", when necessary.

Again, what if the thing using the newer version is not in the wikibase repo (see previous gerrit comment)

Patch-set: 3