Replace assertTag with assertContains in Entity diff classes
All these assertions are super trivial. They do not do much, just
searching for an HTML element by name and if it does have a specific
content. The "exact" same can be done with a substring search for
"<tag>content</tag>". I'm stripping the start tag and searching for
">content</tag>" because the start tag contains attributes.

Note that this patch is meant as an intermediate step to get rid of
assertTag fast. We still want a proper HTML assertion framework for
this.

Bug: T69122
Change-Id: Ia27b83a51da38bb36abcfd676d895a95d9988c72