Update patch set 1
Patch Set 1:

Hmm... for the GET request, ok, but we should really use the master connection on the POST request.

Basically, this change is the equivalent of the Edit page loading article text from a slave DB, increasing the chance of an edit conflict.

Generally, edit should always be based on the latest (master) revision, to avoid conflicts. Are we sure edit conflicts are properly detected if we base the edit on a stale revision? That would require the base revision ID to be passed through the form in a hidden field. I'm not seeing code for that in the base class...

Anyway, what problem does this solve? The queries performed here are pretty trivial.

Patch-set: 1
Label: Code-Review=0