Update patch set 2
Patch Set 2: Code-Review+1

Leszek, are you going to update this? Otherwise I will amend this patch, if this is ok for you.

I also wonder what the reason for this change is? Why is this needed? It does not remove the WikibaseRepo dependency, but only a single service.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1