Update patch set 3
Patch Set 3:

> Wikibase is not a final product, but a building block!
Didn't think it through. Then I agree with Addshore, that version number should be as open as possible.

On the other hand with this approach versions might be corrupted very easily, because developer didn't notice that some library version needs to be updated. Take data-values/geo as an example: composer.json states that minimal version is 1.0, but latest is 1.2 and it will be pulled on developer's machine. When developer writes code he/she sees new methods and just uses them. So now we have that code actually depends on v1.2 but composer.json states that v1.0 is ok. 
I believe that it's a problem, but may be another one.

And what still bothers me is that libraries' versions on developer's machine differ (may be a lot, depending on other extensions' requirements) from versions on production server and it's difficult to be sure if your code works correctly there.

Patch-set: 3
Label: Verified=0