Update patch set 3
Patch Set 3:

I believe there is no value in injecting the WikibaseRepo top level factory. You can as well leave the code as it is. Or am I missing something? What is the benefit of having WikibaseRepo injected?

I hope we will never have code that tries to do something like $this->getMockBuilder( WikibaseRepo::class ). That would be bad. And indeed, we do not have this at the moment.

My proposed solution: Needs a refactoring of the ApiHelperFactory class, so it does not need to know the context on construction time, but instead pulls the context from the provided $apiBase when needed (and only when needed, which happens exclusively in getEntitySavingHelper).

Patch-set: 3