Update patch set 1
Patch Set 1:

I believe I understand what this patch does: An edit may be based on a stale revision, which is usually fine with the diffing and patching we do. But it increases the risk of an edit conflict.

But I wonder: To me this looks like the second most important place in our entire code base that should use a master connection. (The top one being the EditEntity API module.) If not even *this* needs a master connection, why do we even have support for using master connections?

Patch-set: 1
Label: Verified=0