Update patch set 2
Patch Set 2:

Whoops. I believe Aude is right, Jonas. I believe there is no benefit in having a data type that allows a mixture of tabular data and geo shapes. That would be confusing for consumers. It should be either or.

This also solves my naming problem then: "CommonsTabularDataType" and "CommonsMapDataType" (because of the .tab and .map file extensions actually used).

Patch-set: 2